DSS-3165 (#25)
For this ticket:
https://timeoutgroup.atlassian.net/browse/DSS-3165?atlOrigin=eyJpIjoiNzc1OGI1OWJiODljNDA1YzlkMGVkMjlkMDRlMGQyMzgiLCJwIjoiaiJ9
I have looked into updating to-data-library to include a process for
uploading files from s3 to big query, however, when looking into the
library I found the method already exists as well as a method to
transfer files from s3 to gcs.

I tested using the methods in an existing splash da and found the way
partitioning was defined was not compatible with our code so I have made
changes to correct this and keep the code similar to the gs_to_bq method
(which I have used before).

I also have added some schema validation and tests. I would be
interested to know if you think we need more schema validation and if
some of the tests are too repetitive.
Also do you think the schema structure is sufficient? I have tried to
look at the data for the ads reporting but not sure I'm looking in the
right place.