DSS-3231 Add success return to gs_to_bq transfer method (#31)
This is because it can cause silent failures without it. There is a
try-except (without a 'raise') within the load_table_from_uris method,
so if there is an error in the BigQuery load job, the try-except causes
it to appear as a success. Hence we can't then ascertain whether the job
failed and raise an error accordingly.

See JIRA ticket comment
[here](https://timeoutgroup.atlassian.net/browse/DSS-3231?focusedCommentId=263012)
for local testing of this method (using a copy of the transfer module)